+++ tests/check/libs/audioencoder.c
@@ -283,7 +283,7 @@ GST_START_TEST (audioencoder_flush_events)
 {
   GstSegment segment;
   GstBuffer *buffer;
-  guint64 i;
+  guint i;
   GList *events_iter;
 
   setup_audioencodertester ();
